+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Olympia">
         <activity
-            android:name=".ViewHistory.ViewHistoryWorkout"
+            android:name=".ViewHistory.workoutPlanList"
             android:exported="false" />
         <activity
             android:name=".ViewHistory.ViewHistoryActivity"
