--- app/src/main/java/com/example/olympia/ViewHistory/ViewHistoryWorkout.java
@@ -1,45 +1,0 @@
-package com.example.olympia.ViewHistory;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.fragment.app.FragmentTransaction;
-
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.example.olympia.CalorieCounter.ResultList;
-import com.example.olympia.R;
-import com.github.mikephil.charting.charts.BarChart;
-import com.github.mikephil.charting.data.BarData;
-import com.github.mikephil.charting.data.BarDataSet;
-import com.github.mikephil.charting.data.BarEntry;
-import com.github.mikephil.charting.utils.ColorTemplate;
-
-public class ViewHistoryWorkout extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_view_history);
-
-        //for toolbar
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setTitle("History");
-
-//        Button foodButton = findViewById(R.id.idFoodButton);
-        Button workButton = findViewById(R.id.idFoodButton);
-
-        workButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //fragmentTransaction - Start Result List
-                FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
-                ft.replace(R.id.idPlaceHolder, new WorkoutGraphFragment());
-                ft.commit();
-            }
-        });
-    }
-}
