+++ app/src/main/java/com/example/olympia/ViewHistory/workoutPlanList.java
@@ -1,0 +1,28 @@
+package com.example.olympia.ViewHistory;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.FragmentTransaction;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.TextView;
+
+import com.example.olympia.R;
+
+public class workoutPlanList extends AppCompatActivity {
+    // Variables
+    TextView planName, planData;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_view_history);
+
+        //for toolbar
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        getSupportActionBar().setTitle("History");
+    }
+}
