--- app/src/main/res/layout/activity_view_history_workout.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ViewHistory.ViewHistoryWorkout">
-    <include
-        android:id = "@+id/toolbar"
-        layout="@layout/toolbar"
-        />
-    <TextView
-        android:id="@+id/workoutPlanTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Workout Plans"
-        android:textSize="32dp"
-        android:layout_below="@id/toolbar"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="50dp"
-
-        />
-
-    <FrameLayout
-        android:id="@+id/idPlaceHolder"
-        android:layout_width="match_parent"
-        android:layout_height="400dp"
-        android:layout_below="@id/workoutPlanTitle"
-        android:layout_marginTop="50dp" />
-
-    <Button
-        android:id="@+id/idWorkoutButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/idPlaceHolder"
-        android:layout_marginLeft="70dp"
-        android:text="Workout"
-        android:layout_marginTop="50dp"
-        />
-
-    <Button
-        android:id="@+id/idFoodButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/idPlaceHolder"
-        android:layout_marginLeft="60dp"
-        android:layout_marginTop="50dp"
-        android:layout_toRightOf="@id/idWorkoutButton"
-        android:text="Food" />
-
-
-
-
-
-</RelativeLayout>
