+++ app/src/main/res/layout/activity_view_history_workout_plan_list.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ViewHistory.workoutPlanList">
+    <include
+        android:id = "@+id/toolbar"
+        layout="@layout/toolbar"
+        />
+    <TextView
+        android:id="@+id/workoutPlanTitle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Workout Plans"
+        android:textSize="32dp"
+        android:layout_below="@id/toolbar"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="50dp"
+
+        />
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/planCard1"
+        android:layout_width="wrap_content"
+        android:layout_height="102dp"
+        android:layout_below="@id/workoutPlanTitle"
+        android:layout_centerHorizontal="true"
+        android:layout_marginLeft="150dp"
+        android:layout_marginTop="30dp"
+        app:cardBackgroundColor="#12d4e6"
+        app:cardCornerRadius="20dp">
+
+        <TextView
+            android:id="@+id/planName"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="10dp"
+            android:layout_marginTop="10dp"
+            android:text="TextView"
+            android:textSize="25dp" />
+
+        <TextView
+            android:id="@+id/planDate"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="right"
+            android:layout_marginTop="10dp"
+            android:layout_marginRight="10dp"
+            android:text="TextView" />
+
+        <FrameLayout
+            android:id="@+id/exerciseListFrame"
+            android:name="com.example.olympia.Exercises.LogExercise.FragmentLogButtonList"
+            android:layout_width="244dp"
+            android:layout_height="45dp"
+            android:layout_gravity="bottom"
+            android:layout_marginLeft="10dp"
+            android:layout_marginBottom="10dp">
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/idCircleList"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent" />
+
+        </FrameLayout>
+
+
+    </androidx.cardview.widget.CardView>
+
+
+
+
+
+</RelativeLayout>
